backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
numpy==1.26.2
scipy==1.11.4
opencv-python-headless==4.8.1.78
scikit-image==0.22.0
